Remove fetching wallet rewards in home screen and revert it back to Profile's job
